requirements.txt
Flask==2.0.3
tensorflow
opencv-python-headless==4.5.5.62
spotipy==2.19.0
numpy==1.21.2
python-dotenv==0.19.1
